requirements.txt
kafka-python==2.0.2
psycopg[binary]==3.1.18
sentence-transformers==2.2.2
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
